app/utils/jwt_util.py
# app/utils/jwt_util.py
from datetime import datetime, timedelta
from jose import jwt, JWTError
from app.core.settings import settings

# 生成token
def create_access_token(data: dict):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(minutes=settings.JWT_EXPIRE_MINUTES)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(
        to_encode, settings.JWT_SECRET_KEY, algorithm=settings.JWT_ALGORITHM
    )
    return encoded_jwt

# 校验token
def verify_access_token(token: str):
    try:
        payload = jwt.decode(
            token, settings.JWT_SECRET_KEY, algorithms=[settings.JWT_ALGORITHM]
        )
        user_id: int = int(payload.get("sub"))
        if user_id is None:
            raise ValueError("token无用户ID")
        return user_id
    except JWTError:
        raise ValueError("token无效/过期")