pyproject.toml
[tool.poetry]
name = "py_backend"
version = "0.1.0"
description = "Python Backend Project"
authors = ["Your Name <you@example.com>"]
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.128.7"
uvicorn = "^0.40.0"
sqlalchemy = "^2.0.46"
redis = "^7.1.1"
pyjwt = "^2.11.0"
python-dotenv = "^1.2.1"
pydantic-settings = "^2.12.0"
psycopg2-binary = "^2.9.11"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-v"